backendside/requirements/base.txt
Django==3.1.3
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-restful-admin==1.1.3
djangorestframework-csv==2.1.0
django-filter==2.4.0
django-admin-autocomplete-filter==0.6.1
django-admin-interface==0.14.0
django-admin-list-filter-dropdown==1.0.3
django-admin-numeric-filter==0.1.6
django-admin-rangefilter==0.6.3
django-cors-headers==3.5.0
django-tinymce==3.3.0
django-qr-code==2.1.0
django-mysql==3.9
django-multipleselectfield==0.2.0.post2
django-environ==0.4.5
django-webpack-loader==0.7.0
mysqlclient==2.0.1
requests==2.22.0
requests-oauthlib==1.2.0
certifi==2019.6.16
pika==1.1.0
Pillow>=8.0